비디오 오너가 아닌 경우 접근 막기 403(forbidden)

+++ src/controllers/videoController.js
@@ -25,22 +25,32 @@ export const watch = async (req, res) => {
 
 export const getEdit = async (req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
     const video = await Video.findById(id);
 
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found"});
     }
 
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     return res.render("edit", { pageTitle: `Editing: ${video.title}`, video });
 }
 
 export const postEdit = async (req, res) => {
+    const { user: {_id} } = req.session;
     const { id } = req.params; // const id = req.params.id;
     const {title, description, hashtags } = req.body;
     const video = await Video.exists({ _id: id });
 
     if (!video) {
         return res.status(404).render("404", { pageTitle: "Video not found"});
+    }
+
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
     }
 
     await Video.findByIdAndUpdate(video, {
@@ -80,6 +90,17 @@ export const postUpload = async (req, res) => {
 
 export const deleteVideo = async(req, res) => {
     const { id } = req.params;
+    const { user: {_id} } = req.session;
+    const video = await Video.findById(id);
+
+    if (!video) {
+        return res.status(404).render("404", { pageTitle: "Video not found"});
+    }
+
+    if (String(video.owner) !== String(_id)) {
+        return res.status(403).redirect("/");
+    }
+
     await Video.findByIdAndDelete(id);
     // delete video
     return res.redirect("/");
